Create README.md
When you select this form, 
select the "Refresh" Button, this will load the combo boxes, Category, Curr QTY, and To Buy

When you select an item from the combo box, curr Qty should populate with the quantity your storage has on hand, and if you are in need of buying
anything from storage, the "To Buy" will show a suggested amount to buy of the item selected. If the item you select does not need to be purchased, nothing
will populate in the "To Buy" section.

after you enter your already purchased QTY to enter into your storage stock then select the "Save Item" button. 

The other buttons are "Generate Shopping list" which will basically do that, create a shopping list of items needed for storage.
and Exit, which is explanatory.

Then there is "Donations Aggregate" which sums up all donations and organization donated to and gives you the total for the year.
and then View EOY Aggregate, this gives you the entire year of income and expenditures. this will enter all data to column B with the header Total and the year.
This will give you each consecutive year only if the previous header does not have the current year. To test this change "EOY 2025 Totals" to "EOY 2024 Totals"
then execute the EOY Aggregate selection again, then "EOY 2025 Totals" should populate in column C.

Also, on dashboard there are buttons you can select to view individual expenditures.

The Sheet "ItemList" can be changed/Updated with different Items for storage Column A and the donations category can be changed/Updated Column H. Any updates you make will update the selection combo boxes.

I think I covered it all. Please let me know if you have an idea to make it better or if errors occur.